pre-release v0.2

+++ sketchy/terminal/predict/commands.py
@@ -3,50 +3,62 @@ import shutil
 
 from pathlib import Path
 
-from sketchy.sketchy import Sketchy
+from sketchy.minhash import MashScore
 
 
 @click.command()
 @click.option(
-    '--fastq', '-f', help='Input FASTQ file to predict from.'
+    '--fastq', '-f', help='Input FASTQ file to predict from.', type=Path,
 )
 @click.option(
-    '--sketch', '-s', default='saureus-v1.msh', help='MASH sketch to query'
+    '--sketch', '-s', default='saureus-v1.msh', help='MASH sketch to query', type=Path
 )
 @click.option(
-    '--data', '-d', help='Index data file to pull genotypes and other data from.'
+    '--data', '-d', help='Index data file for pull genotypes.', type=Path
 )
 @click.option(
-    '--tmp', '-t', default=Path().home() / '.sketchy' / 'tmp',
-    help='Temporary dir for slicing Fastq file'
+    '--reads', '-r', default=500, help='Number of reads to type.', type=int
+)
+@click.option(
+    '--tmp', '-t', default=Path().cwd() / 'tmp', type=Path,
+    help='Temporary directory for slicing read file.'
+)
+@click.option(
+    '--keep', '-k', is_flag=True,
+    help='Keep temporary folder with match tables for each read / read set.'
 )
 @click.option(
     '--cores', '-c', default=8, help='Number of processors for `mash dist`'
 )
 @click.option(
-    '--header', '-h', is_flag=True, help='Print header to online mode STDOUT.'
+    '--mode', type=str, default="single",
+    help='Analysis mode; single, cumulative, direct.'
 )
 @click.option(
-    '--reads', '-r', default=50, help='Number of reads to type.'
-)
-@click.option(
-    '--single', is_flag=True, help='Single read analysis, instead of cumulative, does not compute score'
-)
-@click.option(
-    '--dist', is_flag=True, help='Use best hits from min hash distance, instead shared hashes.'
+    '--dist', is_flag=True,
+    help='Use smallest MinHash distance, instead of most shared hashes.'
 )
 @click.option(
     '--output', '-o', default=Path().cwd() / 'shared_hashes.csv',
-    help='Top shared hash queries per read, not cumulative, does not compute scores.'
+    help='Output summary files, use --keep to'
 )
-def predict(fastq, sketch, data, tmp, cores, header, reads, single, output, dist):
+def predict(
+        fastq,
+        sketch,
+        data,
+        tmp,
+        keep,
+        cores,
+        reads,
+        mode,
+        output,
+        dist
+):
 
     """ Online lineage matching from uncorrected nanopore reads"""
 
-    sketchy = Sketchy()
-
-    fastq_path = Path(fastq).resolve()
-    sketch_path = Path(sketch).resolve()
+    fastq_path = Path(fastq)
+    sketch_path = Path(sketch)
 
     if not fastq_path.exists():
         print(f'File {fastq_path} does not exist.')
@@ -56,20 +68,32 @@ def predict(fastq, sketch, data, tmp, cores, header, reads, single, output, dist
         print(f'Mash sketch {sketch_path} does not exist.')
         exit(1)
 
+    tmp.mkdir(parents=True, exist_ok=True)
+
     try:
-        sketchy.predict_nanopore(
+
+        ms = MashScore()
+        _ = ms.run(
             fastq=fastq_path,
-            sketch=sketch_path,
-            data=data,
-            tmp=Path(tmp).resolve(),
-            cores=cores,
-            header=header,
             nreads=reads,
-            score=not single,
+            sketch=sketch_path,
+            cores=cores,
+            top=10,
+            mode=mode,
+            data=data,
             out=output,
-            sort_by='dist' if dist else 'shared'
+            sort_by='dist' if dist else 'shared',
+            tmpdir=tmp,
+            online=False
         )
+
     except KeyboardInterrupt:
-        shutil.rmtree(tmp)
+        if not keep:
+            shutil.rmtree(tmp)
+    except AttributeError:
+        # KeyboardInterrupt SIGKILL to running Popen.PIPE
+        if not keep:
+            shutil.rmtree(tmp)
     else:
-        shutil.rmtree(tmp)
+        if not keep:
+            shutil.rmtree(tmp)
